chore: Version Packages (#5965)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to develop, this PR
will be updated.


# Releases
## @pancakeswap/wagmi@0.1.0

### Minor Changes

-   0205708b4: Add Trust Wallet Connector #5932

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>