chore: Version Packages (#5992)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to develop, this PR
will be updated.


# Releases
## @pancakeswap/awgmi@0.0.11

### Patch Changes

-   e68ad6524: Add missing chains on connect

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>